requirements.txt
accelerate==1.12.0
agno==2.4.7
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
asgiref==3.11.0
attrs==25.4.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorlog==6.10.1
cryptography==46.0.4
cuda-bindings==12.9.4
cuda-pathfinder==1.3.3
deprecation==2.1.0
dill==0.4.1
distro==1.9.0
Django==6.0.1
django-picklefield==3.4.0
django-q2==1.9.0
docling==2.71.0
docling-core==2.62.0
docling-ibm-models==3.11.0
docling-parse==4.7.3
docstring_parser==0.17.0
et_xmlfile==2.0.0
Faker==40.1.2
filelock==3.20.3
filetype==1.2.0
fsspec==2026.1.0
gitdb==4.0.12
GitPython==3.1.46
google-api-core==2.29.0
google-api-python-client==2.189.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
huggingface_hub==0.36.1
hyperframe==6.1.0
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lance-namespace==0.4.5
lance-namespace-urllib3-client==0.4.5
lancedb==0.27.1
langchain-core==1.2.8
langchain-openai==1.1.7
langsmith==0.6.8
latex2mathml==3.78.1
lxml==6.0.2
Markdown==3.10.1
markdown-it-py==4.0.0
marko==2.2.2
MarkupSafe==3.0.3
martor==1.7.16
mdurl==0.1.2
mpire==2.10.2
mpmath==1.3.0
multiprocess==0.70.19
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
omegaconf==2.3.0
openai==2.16.0
opencv-python==4.13.0.90
openpyxl==3.1.5
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pillow==11.3.0
pluggy==1.6.0
polyfactory==3.2.0
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pyclipper==1.4.0
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pylatexenc==2.10
pyparsing==3.3.2
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
rapidocr==3.6.0
referencing==0.37.0
regex==2026.1.15
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.3.2
rpds-py==0.30.0
rsa==4.9.1
rtree==1.4.1
safetensors==0.7.0
scipy==1.17.0
semchunk==2.2.2
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.5
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==4.57.6
tree-sitter==0.25.2
tree-sitter-c==0.24.1
tree-sitter-javascript==0.25.0
tree-sitter-python==0.25.0
tree-sitter-typescript==0.23.2
typer==0.21.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.14.0
webencodings==0.5.1
xlsxwriter==3.2.9
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
